requirements.txt
fastapi
uvicorn
requests
google-generativeai
googletrans==4.0.0-rc1
pydantic
